Refactor apis/run to standalone go module
