+++ apis/run/config/crd/bases/run.tanzu.vmware.com_kappcontrollerconfigs.yaml
@@ -1,0 +1,170 @@
+---
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  annotations:
+    controller-gen.kubebuilder.io/version: v0.8.0
+  creationTimestamp: null
+  name: kappcontrollerconfigs.run.tanzu.vmware.com
+spec:
+  group: run.tanzu.vmware.com
+  names:
+    kind: KappControllerConfig
+    listKind: KappControllerConfigList
+    plural: kappcontrollerconfigs
+    shortNames:
+    - kappconf
+    singular: kappcontrollerconfig
+  scope: Namespaced
+  versions:
+  - additionalPrinterColumns:
+    - description: The namespace in which kapp-controller is deployed
+      jsonPath: .spec.namespace
+      name: Namespace
+      type: string
+    - description: The namespace value used for global packaging resources. Any Package
+        and PackageMetadata CRs within that namespace will be included in all other
+        namespaces on the cluster, without duplicating them
+      jsonPath: .spec.kappController.globalNamespace
+      name: GlobalNamespace
+      type: string
+    - description: Name of the kapp-controller data values secret
+      jsonPath: .status.secretName
+      name: SecretName
+      type: string
+    name: v1alpha3
+    schema:
+      openAPIV3Schema:
+        description: KappControllerConfig is the Schema for the kappcontrollerconfigs
+          API
+        properties:
+          apiVersion:
+            description: 'APIVersion defines the versioned schema of this representation
+              of an object. Servers should convert recognized schemas to the latest
+              internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
+            type: string
+          kind:
+            description: 'Kind is a string value representing the REST resource this
+              object represents. Servers may infer this from the endpoint the client
+              submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
+            type: string
+          metadata:
+            type: object
+          spec:
+            description: KappControllerConfigSpec defines the desired state of KappControllerConfig
+            properties:
+              kappController:
+                default:
+                  deployment:
+                    hostNetwork: true
+                properties:
+                  config:
+                    properties:
+                      caCerts:
+                        description: A cert chain of trusted CA certs. These will
+                          be added to the system-wide cert pool of trusted CA's. Cluster-wide
+                          CA Certificate setting will be used if this is not provided.
+                        type: string
+                      dangerousSkipTLSVerify:
+                        description: A comma delimited list of hostnames for which
+                          kapp-controller should skip TLS verification
+                        type: string
+                      httpProxy:
+                        description: The url/ip of a proxy for kapp controller to
+                          use when making network requests. Cluster-wide HTTP proxy
+                          setting will be used if this is not provided.
+                        type: string
+                      httpsProxy:
+                        description: The url/ip of a TLS capable proxy for kapp-controller
+                          to use when making network requests. Cluster-wide HTTPS
+                          proxy setting will be used if this is not provided.
+                        type: string
+                      noProxy:
+                        description: A comma delimited list of domain names which
+                          kapp-controller should bypass the proxy for when making
+                          requests. Cluster-wide no-proxy setting will be used if
+                          this is not provided.
+                        type: string
+                    type: object
+                  createNamespace:
+                    default: false
+                    description: Whether to create namespace specified for kapp-controller
+                    type: boolean
+                  deployment:
+                    default:
+                      hostNetwork: true
+                    properties:
+                      apiPort:
+                        default: 10100
+                        description: Bind port for kapp-controller API
+                        type: integer
+                      concurrency:
+                        default: 4
+                        description: Concurrency of kapp-controller deployment
+                        type: integer
+                      hostNetwork:
+                        default: true
+                        description: Whether to enable host networking for kapp-controller
+                          deployment
+                        type: boolean
+                      metricsBindAddress:
+                        default: "0"
+                        description: Address for metrics server
+                        type: string
+                      priorityClassName:
+                        default: system-cluster-critical
+                        description: The priority value that various system components
+                          use to find the priority of the kapp-controller pod
+                        type: string
+                      tolerations:
+                        default:
+                        - key: CriticalAddonsOnly
+                          operator: Exists
+                        - effect: NoSchedule
+                          key: node-role.kubernetes.io/master
+                        - effect: NoSchedule
+                          key: node.kubernetes.io/not-ready
+                        - effect: NoSchedule
+                          key: node.cloudprovider.kubernetes.io/uninitialized
+                          value: "true"
+                        description: kapp-controller deployment tolerations
+                        items:
+                          additionalProperties:
+                            type: string
+                          type: object
+                        type: array
+                    type: object
+                  globalNamespace:
+                    default: tkg-system
+                    description: The namespace value used for global packaging resources.
+                      Any Package and PackageMetadata CRs within that namespace will
+                      be included in all other namespaces on the cluster, without
+                      duplicating them
+                    type: string
+                type: object
+              namespace:
+                default: tkg-system
+                description: The namespace in which kapp-controller is deployed
+                type: string
+            type: object
+          status:
+            description: KappControllerConfigStatus defines the observed state of
+              KappControllerConfig
+            properties:
+              secretRef:
+                description: Name of the data value secret created by controller
+                type: string
+            type: object
+        required:
+        - spec
+        type: object
+    served: true
+    storage: true
+    subresources:
+      status: {}
+status:
+  acceptedNames:
+    kind: ""
+    plural: ""
+  conditions: []
+  storedVersions: []
