+++ apis/run/v1alpha1/conversion.go
@@ -8,9 +8,9 @@ import (
 	apiconversion "k8s.io/apimachinery/pkg/conversion"
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 
-	v1alpha3 "github.com/vmware-tanzu/tanzu-framework/apis/run/v1alpha3"
-	"github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/util/version"
-	utilconversion "github.com/vmware-tanzu/tanzu-framework/util/conversion"
+	utilconversion "github.com/vmware-tanzu/tanzu-framework/apis/run/util/conversion"
+	"github.com/vmware-tanzu/tanzu-framework/apis/run/util/version"
+	"github.com/vmware-tanzu/tanzu-framework/apis/run/v1alpha3"
 )
 
 const (
