+++ pkg/v1/tkg/test/tkgctl/shared/e2e_tkrresolver.go
@@ -22,6 +22,7 @@ import (
 	"sigs.k8s.io/cluster-api/util/conditions"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/vmware-tanzu/tanzu-framework/apis/run/util/version"
 	runv1 "github.com/vmware-tanzu/tanzu-framework/apis/run/v1alpha3"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/constants"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/test/framework"
@@ -32,7 +33,6 @@ import (
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/util/resolution"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/util/sets"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/util/topology"
-	"github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/util/version"
 	clusterdata "github.com/vmware-tanzu/tanzu-framework/pkg/v2/tkr/webhook/cluster/tkr-resolver/cluster"
 )
